+++ Tests/Microsoft.AppCenter.Test.WindowsDesktop.Shared/Utils/ApplicationSettingsTest.cs
@@ -1,17 +1,17 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.AppCenter.Utils;
 using System;
 using System.IO;
 using System.Reflection;
-using Microsoft.AppCenter.Utils;
 using Xunit;
 
 namespace Microsoft.AppCenter.Test.WindowsDesktop.Utils
 {
     public class ApplicationSettingsTest : IDisposable
     {
-        private readonly DefaultApplicationSettings _settings;
+        private DefaultApplicationSettings _settings;
 
         public ApplicationSettingsTest()
         {
@@ -20,7 +20,13 @@ namespace Microsoft.AppCenter.Test.WindowsDesktop.Utils
 
         public void Dispose()
         {
-            File.Delete(DefaultApplicationSettings.FilePath);
+            try
+            {
+                File.Delete(DefaultApplicationSettings.FilePath);
+            }
+            catch (Exception)
+            {
+            }
         }
 
         /// <summary>
@@ -62,6 +68,43 @@ namespace Microsoft.AppCenter.Test.WindowsDesktop.Utils
             Assert.Equal(42, _settings.GetValue<int>(key));
             _settings.Remove(key);
             Assert.False(_settings.ContainsKey(key));
+        }
+
+        [Fact]
+        public void VerifyMigration()
+        {
+            // Set a value.
+            _settings.SetValue("old", "value");
+
+            // Move file back to old location to test migration.
+            var oldLocation = typeof(DefaultApplicationSettings).Assembly.Location;
+            var oldPath = Path.Combine(Path.GetDirectoryName(oldLocation), "AppCenter.config");
+            File.Move(DefaultApplicationSettings.FilePath, oldPath);
+
+            // Migrate.
+            _settings = new DefaultApplicationSettings();
+
+            // Check.
+            Assert.Equal("value", _settings.GetValue<string>("old"));
+        }
+
+        [Fact]
+        public void VerifyMigrationSkippedWhenNewFileExists()
+        {
+            // Set a value.
+            _settings.SetValue("key", "oldValue");
+
+            // Move file back to old location to test migration.
+            var oldLocation = typeof(DefaultApplicationSettings).Assembly.Location;
+            var oldPath = Path.Combine(Path.GetDirectoryName(oldLocation), "AppCenter.config");
+            File.Copy(DefaultApplicationSettings.FilePath, oldPath);
+            _settings.SetValue("key", "newValue");
+
+            // Migrate.
+            _settings = new DefaultApplicationSettings();
+
+            // Check migration didn't happen.
+            Assert.Equal("newValue", _settings.GetValue<string>("key"));
         }
     }
 }