+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 This preview version adds support for .NET Core 3.0 WPF and WinForms applications for the Analytics and Crashes modules.
 
+## App Center
+
+#### WPF/WinForms
+
+* **[Fix]** Fix configuration file location to be in a user-specific directory.
+
 ### App Center Analytics
 
 #### Windows